refactor: keep quarantine validation files within size gate
